Added a project "repository" (#28)
The core mechanism for the engagement score calculator is to scan all issues in a project and calculate some number that represents the amount of "engagement" that is happening on a particular issue.

This PR adds the support for a "project" concept so we can query for items, issues and maybe more in the future. It integrates with the existing repository to fetch comments.